+++ python/ifigure/add_on/model/module/subs/dwscope_subs.py
@@ -79,12 +79,12 @@ def import_dwscope(file):
 
     f.close()
 
-    num_c = long(d["columns"])
+    num_c = int(d["columns"])
     num_r = [0]*num_c
     return d
     for i in range(num_c)+1:
         key = 'rows_in_column_'+str(i)
-        num_r[i-1] = long(d[key])
+        num_r[i-1] = int(d[key])
 
     d2 = {"num_c": num_c,
           "num_r": num_r}
