+++ python/ifigure/mto/fig_image.py
@@ -630,11 +630,11 @@ class FigImage(FigObj, XUser, YUser, ZUser, CUser):
             dy = np.floor(p1[1]-p0[1])+2
             xp = idtrans(
                 np.transpose(
-                    np.vstack((np.floor(p0[0])+np.arange(long(dx)),
+                    np.vstack((np.floor(p0[0])+np.arange(int(dx)),
                                np.linspace(p0[0], p1[0], dx)))))[:, 0]
             yp = idtrans(
                 np.transpose(
-                    np.vstack((np.floor(p0[1])-1+np.zeros(long(dy)),
+                    np.vstack((np.floor(p0[1])-1+np.zeros(int(dy)),
                                np.linspace(p0[1], p1[1], dy)))))[:, 1]
 
             # eliminate points outside the data range
