+++ python/ifigure/rpo/proxy_nete.py
@@ -8,7 +8,7 @@ class ProxyNeTe(object):
     def fill_array(self, i, txt):
         ct = []
         print((i, txt[i]))
-        for t in range(1+i, long(txt[i])+1+i):
+        for t in range(1+i, int(txt[i])+1+i):
             ct.append(txt[t])
         return t+1, ct
